Correcciones 08-05
